Refactored Tracks view, Edit Track view, WMS Layers view
Renamed cteco.list.html and cteco.list.controller.js to wms.list.html
and wms.list.controller.js to reflect changes to the views. Renamed CT
ECO view to WMS Layers, which includes the CT ECO, ortho, and other WMS
layers that can be toggled on/off. Added ionicons to the menu and
icomoon icons to some buttons in the Tracks views.